requirements.txt
Django==1.7.1
httplib2==0.8
restclient==0.11.0
feedparser==4.1
Markdown==2.3.1
pyenchant==1.5.1
smartypants==1.6.0.2
uuid==1.30
psycopg2==2.5.1
Pillow==2.3.0
statsd==2.0.2
nose==1.1.2
coverage==3.5.3
python-ldap==2.4.13
pyflakes==0.5.0
pep8==1.4
flake8==1.7.0
logilab-common==0.59.0
logilab-astng==0.24.1
pylint==0.26.0

django-courseaffils==1.0.0
django-smartif==0.1
djangowind==0.13.2
sorl==3.1
template_utils==0.4p2
typogrify==1.0
pagetree==0.5.9
django-pageblocks==1.0.0
djangohelpers==0.2
django-statsd-mozilla==0.3.11
django-nose==0.1.3
django-jenkins==0.16.3
django-smoketest==0.3.0
raven==4.0.3
factory_boy==2.1.1
django-waffle==0.9.0
django-annoying==0.8.0
django-markwhat==1.0
